Move CFolderProcessingThread out of the queue into own file and begin separating the logic.
The goal is to have adding local directories to the queue behave exactly as adding remote directories to the queue, with the recursive operation status below the local file list. Under the hood, both local and remote recursive operations should have the same interface.

This common functionality can then at a later step be used to have a local file search and possibly even a comparative search.

git-svn-id: https://svn.filezilla-project.org/svn/FileZilla3/trunk@7363 ce1b9ea0-41a8-11dc-9b12-cb80f3b22977