Fix type mismatch in sftp_find_request
The id member of the sftp_request structure is of type unsigned int. This type is also used in the sftp_reqfind callback. In sftp_find_request we thus need to pass a pointer to unsigned int to find234. Before this commit, sftp_find_request was passing a pointer to unsigned long to find234, which causes the lookup to fail on big-endian platforms where sizeof(unsigned int) != sizeof(unsigned long), e.g. ppc64.

git-svn-id: https://svn.filezilla-project.org/svn/FileZilla3/trunk@8002 ce1b9ea0-41a8-11dc-9b12-cb80f3b22977